+++ lib/Charge.php
@@ -36,4 +36,17 @@ class Charge extends ApiResource
     {
         return self::_create($params, $options);
     }
+
+    /**
+     * @param string $id The ID of the charge.
+     * @param array|string|null $options
+     *
+     * @return Charge The reversed charge.
+     */
+    public static function reverse($id, $options = null)
+    {
+        $url = static::classUrl().'/'.$id.'/reverse';
+        $params = [];
+        return static::_directRequest('post', $url, $params, $options);
+    }
 }